Also export convertFromRawToContentBlock
